Made accuracy improvements

+++ src/bullet.rs
@@ -14,7 +14,7 @@ use std::f32::consts::PI;
 
 pub struct Bullet {
     pub shooter_id: u32,
-    pub diameter: f32,
+    pub diameter: f64,
     pub body_handle: DefaultBodyHandle,
     pub collider_handle: DefaultColliderHandle,
 }
@@ -49,18 +49,18 @@ impl Bullet {
         bullet_batch: &mut ggez::graphics::spritebatch::SpriteBatch,
         bullet_image_width: f32,
         bullet_image_height: f32,
-        colliders: &mut DefaultColliderSet<f32>,
+        colliders: &mut DefaultColliderSet<f64>,
     ) -> ggez::GameResult {
         let collider_optional = colliders.get(self.collider_handle);
         if let Some(collider) = collider_optional {
-            let position: &Isometry2<f32> = collider.position();
+            let position: &Isometry2<f64> = collider.position();
             let d = graphics::DrawParam::new()
-                .dest([position.translation.x, position.translation.y])
+                .dest([position.translation.x as f32, position.translation.y as f32])
                 .offset([0.5, 0.5])
-                .rotation(position.rotation.re)
+                .rotation(position.rotation.re as f32)
                 .scale([
-                    self.diameter / bullet_image_width,
-                    self.diameter / bullet_image_height,
+                    self.diameter as f32 / bullet_image_width,
+                    self.diameter as f32 / bullet_image_height,
                 ]);
             bullet_batch.add(d);
         }
